backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-jsonform==2.22.0
django-polymorphic==3.1
django-rest-polymorphic==0.1.10
django-smart-selects==1.6.0
django-svg-image-form-field==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.1
drf-nested-routers==0.93.4
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
mpmath==1.3.0
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.21
pydot==1.4.2
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
